Move formatPreviousPeriodOptionName into trend-helpers
trend-helpers imported it from SmartScalar/utils while SmartScalar
imports trend-helpers, so the two files imported each other. The
function only depends on the date unit, so it moves down and
SmartScalar imports it from there.
